refactor(sdk): :recycle: 对齐 OpenAPI/JS SDK(v2) 类型定义与 API 契约
- 整数语义字段 float64 → int64 类型修正（时间戳、行号、计数等）
- 合并分离的 body/query 参数为单一 params 结构体（Mcp、Project、Provider）
- 补充缺失字段与新增字段对齐规范（Path.Home、ProviderInfo、McpOAuthConfig 等）
- 移除冗余类型与清理废弃结构体（SymbolLocationRangeEnd/Start、ProjectUpdateParamsQuery 等）

+++ file.go
@@ -236,10 +236,10 @@ func (r fileReadResponsePatchJSON) RawJSON() string {
 
 type FileReadResponsePatchHunk struct {
 	Lines    []string                      `json:"lines,required"`
-	NewLines float64                       `json:"newLines,required"`
-	NewStart float64                       `json:"newStart,required"`
-	OldLines float64                       `json:"oldLines,required"`
-	OldStart float64                       `json:"oldStart,required"`
+	NewLines int64                         `json:"newLines,required"`
+	NewStart int64                         `json:"newStart,required"`
+	OldLines int64                         `json:"oldLines,required"`
+	OldStart int64                         `json:"oldStart,required"`
 	JSON     fileReadResponsePatchHunkJSON `json:"-"`
 }
